requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
eth-abi>=4.0.0
eth-utils>=2.1.0 